JavaFX Hello World - Scraper portion started
I have a very basic JavaFX application that pops up on the screen with a
button and a text field.  This part is turning out harder than I
expected..I'll have to sit down and spend some time reading through the
documentation to really grasp the different layout options and other
things JavaFX has to choose from.

I also started the scraper portion.  As of now, it takes the initial
link, grabs and stores all other links off that page into an ArrayList,
and stores all words on that page into a separate ArrayList.